requirements.txt
aiogram==3.13.1
aiohttp==3.10.11
python-dotenv==1.1.1
openai-whisper==20250625
torch==2.8.0
torchaudio==2.8.0
chromadb==0.5.0
sentence-transformers==2.2.2